requirements.txt
opencv-python
open3d
numpy
pyyaml
pandas
matplotlib